+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"/Users/pratikyadav/projects/web-app-client/src/components/Slider/Slider.js":"1"},{"size":3055,"mtime":1712833711477,"results":"2","hashOfConfig":"3"},{"filePath":"4","messages":"5","suppressedMessages":"6","errorCount":0,"fatalErrorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"kgt3dg","/Users/pratikyadav/projects/web-app-client/src/components/Slider/Slider.js",[],[]]
