+++ src/components/CollegeResult.css
