+++ src/components/CollegeResult.tsx
