+++ src/components/Navbar/Navbar.tsx
@@ -9,9 +9,9 @@ import React, { useEffect, useState } from 'react'
 import Image from 'next/image'
 //import { IoMdMail } from 'react-icons/io'
 import Link from 'next/link'
-//import demopic from '../../../assets/demopic.jpg'
-import star from '../../../assets/star.png'
-import unibuzzLogo from '../../../assets/logo.svg'
+//import demopic from '@assets/demopic.jpg'
+import star from '@assets/star.png'
+import unibuzzLogo from '@assets/logo.svg'
 import { usePathname } from 'next/navigation'
 import { menuContent } from './constant'
 import { motion } from 'framer-motion'
