+++ src/components/SearchHistoryBox/SearchHistoryBox.css
