+++ src/components/SearchHistoryBox/SearchHistoryBox.tsx
