+++ src/components/Section-2/Section2.css
@@ -4,7 +4,7 @@
   display: flex;
   flex-direction: column;
   justify-content: center;
-  background-image: url('../../../assets/test.png');
+  background-image: url('../../assets/test.png');
   background-position: center; /* Center the image */
   background-repeat: no-repeat; /* Do not repeat the image */
   background-size: cover;
