+++ tsconfig.json
@@ -22,8 +22,10 @@
     ],
     "paths": {
       "@/*": [
-        "./src/*"
-      ]
+        "./src/*",
+      ],
+      "@assets/*": ["./src/assets/*"],
+      "@components/*": ["./src/components/*"]
     },
     "skipLibCheck": true
   },
